Some restructuring and start of auth redirecting
