+++ src/App.js
@@ -6,11 +6,11 @@ import Navbar from "./components/Drawer/drawer"
 import AdminAccessGroups from "./containers/Admin/AdminAccessGroups"
 import AdminRooms from "./containers/Admin/AdminRooms"
 import AdminServices from "./containers/Admin/AdminServices"
-import { Switch, Route } from "react-router-dom"
+import { Switch, Route, Redirect } from "react-router-dom"
 import AdminUsers from "./containers/Admin/AdminUsers"
 import CssBaseline from "@material-ui/core/CssBaseline"
 import { makeStyles } from "@material-ui/core/styles"
-import Bookingsite from "./components/Bookings/BookingSite"
+import Bookingsite from "./containers/BookingContainers/BookingSite"
 import PageNotFound from "./containers/404Page"
 import Dashboard from "./containers/Dashboard/Dashboard.jsx"
 import SignInSide from "./containers/SignIn/SigninSide"
@@ -27,6 +27,26 @@ const useStyles = makeStyles(theme => ({
   },
   appBarSpacer: theme.mixins.toolbar
 }))
+
+function PrivateRoute({ children, ...rest }) {
+  return (
+    <Route
+      {...rest}
+      render={({ location }) =>
+        true ? (
+          children
+        ) : (
+            <Redirect
+              to={{
+                pathname: "/booking",
+                state: { from: location }
+              }}
+            />
+          )
+      }
+    />
+  );
+}
 
 const App = () => {
   const [token, setToken] = React.useState(null)
@@ -47,31 +67,29 @@ const App = () => {
           <div className={classes.appBarSpacer}>
             <Switch>
               <Route exact path="/" render={() => <SignInSide />} />
-              {token && (
-                <>
-                  <Route exact path="/dashboard" render={() => <Dashboard />} />
-                  <Route exact path="/booking" render={() => <Bookingsite />} />
-                  <Route exact path="/admin" render={() => <AdminHome />} />
-                  <Route
-                    exact
-                    path="/services"
-                    render={() => <AdminServices />}
-                  />
-                  <Route exact path="/rooms" render={() => <AdminRooms />} />
-                  <Route
-                    exact
-                    path="/accessgroups"
-                    render={() => <AdminAccessGroups />}
-                  />
-                  <Route exact path="/users" render={() => <AdminUsers />} />
-                </>
-              )}
+              <PrivateRoute path="/dashboard" >
+                <Dashboard />
+              </PrivateRoute>
+              <Route path="/booking" component={Bookingsite} />
+              <Route exact path="/admin" component={AdminHome} />
+              <Route
+                exact
+                path="/services"
+                component={AdminServices}
+              />
+              <Route exact path="/rooms" component={AdminRooms} />
+              <Route
+                exact
+                path="/accessgroups"
+                component={AdminAccessGroups}
+              />
+              <Route exact path="/users" component={AdminUsers} />
               <Route component={PageNotFound} />
             </Switch>
           </div>
         </main>
       </div>
-    </ApolloProvider>
+    </ApolloProvider >
   )
 }
