+++ src/components/Booking/useAddBooking.jsx
@@ -1,6 +1,5 @@
 import gql from "graphql-tag"
 import { useMutation } from "@apollo/react-hooks"
-import { GET_BOOKINGS } from "../Querys/useBookingsByUser"
 
 const CREATE_BOOKING = gql`
   mutation addBooking(
