--- src/components/Bookings/listItems.jsx
@@ -1,65 +1,0 @@
-import React from 'react';
-import ListItem from '@material-ui/core/ListItem';
-import ListItemIcon from '@material-ui/core/ListItemIcon';
-import ListItemText from '@material-ui/core/ListItemText';
-import ListSubheader from '@material-ui/core/ListSubheader';
-import AssignmentIcon from '@material-ui/icons/Assignment';
-import HomeIcon from '@material-ui/icons/Home';
-import BookIcon from '@material-ui/icons/Book';
-import ScheduleIcon from '@material-ui/icons/Schedule';
-import AccountCircleIcon from '@material-ui/icons/AccountCircle';
-import RoomIcon from '@material-ui/icons/Room';
-
-
-export const mainListItems = (
-    <div>
-        <ListItem button>
-            <ListItemIcon >
-                <HomeIcon />
-            </ListItemIcon>
-            <ListItemText primary="Home" />
-        </ListItem>
-        <ListItem button button>
-            <ListItemIcon>
-                <RoomIcon />
-            </ListItemIcon>
-            <ListItemText primary="Rooms" />
-        </ListItem>
-        <ListItem button button>
-            <ListItemIcon>
-                <ScheduleIcon />
-            </ListItemIcon>
-            <ListItemText primary="Book" />
-        </ListItem>
-        <ListItem button button>
-            <ListItemIcon>
-                <AccountCircleIcon />
-            </ListItemIcon>
-            <ListItemText primary="Account" />
-        </ListItem>
-    </div>
-);
-
-export const secondaryListItems = (
-    <div>
-        <ListSubheader inset>Saved reports</ListSubheader>
-        <ListItem button>
-            <ListItemIcon>
-                <AssignmentIcon />
-            </ListItemIcon>
-            <ListItemText primary="Current month" />
-        </ListItem>
-        <ListItem button>
-            <ListItemIcon>
-                <AssignmentIcon />
-            </ListItemIcon>
-            <ListItemText primary="Last quarter" />
-        </ListItem>
-        <ListItem button>
-            <ListItemIcon>
-                <AssignmentIcon />
-            </ListItemIcon>
-            <ListItemText primary="Year-end sale" />
-        </ListItem>
-    </div>
-);
